backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
psycopg2-binary
python-dotenv
